+++ apps/meteor/client/views/omnichannel/hooks/useOmnichannelPriorities.ts
@@ -1,0 +1,3 @@
+import { useOmnichannel } from './useOmnichannel';
+
+export const useOmnichannelPriorities = () => useOmnichannel().livechatPriorities;