+++ apps/meteor/ee/server/methods/getReadReceipts.ts
@@ -1,4 +1,4 @@
-import type { ReadReceipt as ReadReceiptType, IMessage } from '@rocket.chat/core-typings';
+import type { IReadReceiptWithUser, IMessage } from '@rocket.chat/core-typings';
 import type { ServerMethods } from '@rocket.chat/ddp-client';
 import { License } from '@rocket.chat/license';
 import { Messages } from '@rocket.chat/models';
@@ -11,11 +11,11 @@ import { ReadReceipt } from '../lib/message-read-receipt/ReadReceipt';
 declare module '@rocket.chat/ddp-client' {
 	// eslint-disable-next-line @typescript-eslint/naming-convention
 	interface ServerMethods {
-		getReadReceipts(options: { messageId: IMessage['_id'] }): ReadReceiptType[];
+		getReadReceipts(options: { messageId: IMessage['_id'] }): IReadReceiptWithUser[];
 	}
 }
 
-export const getReadReceiptsFunction = async function (messageId: IMessage['_id'], userId: string): Promise<ReadReceiptType[]> {
+export const getReadReceiptsFunction = async function (messageId: IMessage['_id'], userId: string): Promise<IReadReceiptWithUser[]> {
 	if (!License.hasModule('message-read-receipt')) {
 		throw new Meteor.Error('error-action-not-allowed', 'This is an enterprise feature', { method: 'getReadReceipts' });
 	}