+++ apps/meteor/lib/callbacks/beforeAddUserToRoom.ts
@@ -2,5 +2,6 @@ import type { IUser, IRoom } from '@rocket.chat/core-typings';
 
 import { Callbacks } from './callbacksBase';
 
-export const beforeAddUserToRoom =
-	Callbacks.create<(args: { user: IUser; inviter?: Pick<IUser, '_id' | 'username'> }, room: IRoom) => void>('beforeAddUserToRoom');
+export const beforeAddUserToRoom = Callbacks.create<(args: { user: IUser; inviter?: IUser }, room: IRoom) => void>('beforeAddUserToRoom');
+export const beforeAddUsersToRoom =
+	Callbacks.create<(args: { usernames: string[]; inviter?: IUser }, room: IRoom) => void>('beforeAddUsersToRoom');