+++ apps/meteor/tests/e2e/page-objects/home-channel.ts
@@ -1,6 +1,7 @@
 import type { Locator, Page } from '@playwright/test';
 
-import { HomeContent, HomeSidenav, HomeFlextab, Navbar, Sidebar, Sidepanel } from './fragments';
+import { HomeContent, HomeSidenav, HomeFlextab, Navbar, Sidepanel, RoomSidebar } from './fragments';
+import { RoomToolbar } from './fragments/toolbar';
 
 export class HomeChannel {
 	public readonly page: Page;
@@ -9,7 +10,7 @@ export class HomeChannel {
 
 	readonly sidenav: HomeSidenav;
 
-	readonly sidebar: Sidebar;
+	readonly sidebar: RoomSidebar;
 
 	readonly sidepanel: Sidepanel;
 
@@ -17,14 +18,17 @@ export class HomeChannel {
 
 	readonly tabs: HomeFlextab;
 
+	readonly roomToolbar: RoomToolbar;
+
 	constructor(page: Page) {
 		this.page = page;
 		this.content = new HomeContent(page);
 		this.sidenav = new HomeSidenav(page);
-		this.sidebar = new Sidebar(page);
+		this.sidebar = new RoomSidebar(page);
 		this.sidepanel = new Sidepanel(page);
 		this.navbar = new Navbar(page);
 		this.tabs = new HomeFlextab(page);
+		this.roomToolbar = new RoomToolbar(page);
 	}
 
 	goto() {