+++ packages/http-router/src/Router.ts
@@ -1,3 +1,4 @@
+import { Logger } from '@rocket.chat/logger';
 import type { Method } from '@rocket.chat/rest-typings';
 import type { AnySchema } from 'ajv';
 import express from 'express';
@@ -8,6 +9,8 @@ import qs from 'qs'; // Using qs specifically to keep express compatibility
 
 import type { ResponseSchema, TypedOptions } from './definition';
 import { honoAdapterForExpress } from './middlewares/honoAdapterForExpress';
+
+const logger = new Logger('HttpRouter');
 
 type MiddlewareHandlerListAndActionHandler<TOptions extends TypedOptions, TContext = (c: Context) => Promise<ResponseSchema<TOptions>>> = [
 	...MiddlewareHandler[],
@@ -199,6 +202,14 @@ export class Router<
 			if (options.query) {
 				const validatorFn = options.query;
 				if (typeof options.query === 'function' && !validatorFn(queryParams)) {
+					logger.warn({
+						msg: 'Query parameters validation failed - route spec does not match request payload',
+						method: req.method,
+						path: req.url,
+						error: validatorFn.errors?.map((error: any) => error.message).join('\n '),
+						bodyParams: undefined,
+						queryParams,
+					});
 					return c.json(
 						{
 							success: false,
@@ -215,6 +226,14 @@ export class Router<
 			if (options.body) {
 				const validatorFn = options.body;
 				if (typeof options.body === 'function' && !validatorFn((req as any).bodyParams || bodyParams)) {
+					logger.warn({
+						msg: 'Request body validation failed - route spec does not match request payload',
+						method: req.method,
+						path: req.url,
+						error: validatorFn.errors?.map((error: any) => error.message).join('\n '),
+						bodyParams,
+						queryParams: undefined,
+					});
 					return c.json(
 						{
 							success: false,
@@ -240,6 +259,13 @@ export class Router<
 					throw new Error(`Missing response validator for endpoint ${req.method} - ${req.url} with status code ${statusCode}`);
 				}
 				if (responseValidatorFn && !responseValidatorFn(coerceDatesToStrings(body))) {
+					logger.warn({
+						msg: 'Response validation failed - response does not match route spec',
+						method: req.method,
+						path: req.url,
+						error: responseValidatorFn.errors?.map((error: any) => error.message).join('\n '),
+						originalResponse: body,
+					});
 					return c.json(
 						{
 							success: false,
@@ -400,6 +426,14 @@ export class Router<
 			),
 		);
 		return router;
+	}
+
+	getHonoRouter(): Hono<{
+		Variables: {
+			remoteAddress: string;
+		};
+	}> {
+		return this.innerRouter;
 	}
 }